Revert latest changes
This reverts commit daf5a6a3425a4ee7810a53834dec473d41c7d0f3.
