+++ src/components/Results.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { Link, theme, List, ListItem, Button } from "cobra-ui";
+import { Link, theme, List, ListItem } from "cobra-ui";
 import { CopyToClipboard } from "./CopyToClipboard";
 import { ResultsData, ListItemHeader } from "./CommonComponents";
 import {
@@ -24,26 +24,7 @@ const getHumanReadableResult = result => {
   return mapCodeToMessage[result] || "Invalid result";
 };
 
-const downloadFile = file => {
-  const url = URL.createObjectURL(file);
-
-  var tempLink = document.createElement("a");
-  tempLink.style.display = "none";
-  tempLink.href = url;
-  tempLink.setAttribute("download", file.name);
-
-  // Needed because safari thinks _blank anchor are pop ups. We only want to set _blank target if the browser does not support the HTML5 download attribute.
-  if (typeof tempLink.download === "undefined") {
-    tempLink.setAttribute("target", "_blank");
-  }
-
-  document.body.appendChild(tempLink);
-  tempLink.click();
-  document.body.removeChild(tempLink);
-  window.URL.revokeObjectURL(url);
-};
-
-export const AuthenticationResult = ({ files, rawFiles, txtFile }) => (
+export const AuthenticationResult = ({ files }) => (
   <ResultWrapper>
     <ResultLabel>Uploaded digests</ResultLabel>
     <List>
@@ -66,21 +47,6 @@ export const AuthenticationResult = ({ files, rawFiles, txtFile }) => (
                 <span>{file.digest}</span>
                 <CopyToClipboard text={file.digest} />
               </>
-            </ResultsData>
-          </div>
-          <div style={{ display: "flex" }}>
-            <ListItemHeader>Download</ListItemHeader>
-            <ResultsData>
-              <Button
-                small
-                onClick={
-                  txtFile
-                    ? () => downloadFile(txtFile)
-                    : () => downloadFile(rawFiles[i])
-                }
-              >
-                Download
-              </Button>
             </ResultsData>
           </div>
           <div style={{ display: "flex" }}>
@@ -157,7 +123,7 @@ export const VerificationResult = ({ files }) => (
                   })}
                   filename={"digest.json"}
                 >
-                  <Button small>Download</Button>
+                  click here
                 </DownloadFileLink>
               </ResultsData>
             </div>
