--- src/helpers/files.js
@@ -1,15 +1,0 @@
-export const processFiles = files =>
-  new Promise(resolve => {
-    const processedFiles = [];
-    files.forEach(file => {
-      const reader = new FileReader();
-      reader.onload = (f => event => {
-        processedFiles.push({
-          name: f.name,
-          payload: event.target.result.split(",")[1]
-        });
-        if (processedFiles.length === files.length) resolve(processedFiles);
-      })(file);
-      reader.readAsDataURL(file);
-    });
-  });
