+++ src/index.css
@@ -15,19 +15,6 @@ code {
     monospace;
 }
 
-.or {
-  width: 100%;
-  text-align: center;
-  border-bottom: 1px solid #999;
-  line-height: 0.1em;
-  margin: 10px 0 20px;
-}
-
-.or span {
-   background: #fdfdfd;
-   padding: 0 10px;
-}
-
 .lds-ripple {
   display: inline-block;
   position: relative;
